Replaced top_factors.csv with new data set
Built a new data set in test_top_factors using a known decision boundary
and adding some random noise.  The data consists of three useful
variables (two numeric and one categorical) and two variables that are
not useful in the classification.

The data set can be easily modified when top_factors is changed.